+++ model_models_pool.go
@@ -24,7 +24,7 @@ type ModelsPool struct {
 	Passive bool `json:"Passive"`
 	PgAutoscaleMode bool `json:"PgAutoscaleMode"`
 	PgNum int64 `json:"PgNum"`
-	PoolID int32 `json:"PoolID"`
+	PoolID int64 `json:"PoolID"`
 	PoolName string `json:"PoolName"`
 	State string `json:"State"`
 	Status string `json:"Status"`
